qcacld-3.0: Fix to avoid sending invalid PMKID in FT-SAE roam
In FT-SAE when STA connects to AP1 then PMK1 gets cached. And
then STA disconnects from AP1 and connects to AP2 then PMK2
gets cached. Now if STA roams to AP1 then FW uses PMK2 from mlme
session to create the PMKID. But the corresponding RSO command
from driver sends the PMK1 to FW and the same gets updated in
mlme session of FW. This results in failure of next roams, as the
invalid PMKID gets created using PMK1.
This fix helps in deleting the old/stale PMK cache entries that
are under the same mobility domain as of the newly added entry in
case of the FT-SAE connection.

Change-Id: If71b22dbf9200b820c8dd8c9fd0183af5bfcb0f6
CRs-Fixed: 2740532